Actually die if no 'patch' is available, but delay check until when we need it

+++ lib/Devel/PatchPerl.pm
@@ -144,7 +144,6 @@ sub patch_source {
     }
   }
   $source = File::Spec->rel2abs($source);
-  warn "No patch utility found\n" unless $patch_exe;
   {
     my $dir = pushd( $source );
     for my $p ( grep { _is( $_->{perl}, $vers ) } @patch ) {
@@ -182,6 +181,7 @@ sub _patch
   print "patching $_\n" for $patch =~ /^\+{3}\s+(\S+)/gm;
   my $diff = 'tmp.diff';
   _write_or_die($diff, $patch);
+  die "No patch utility found\n" unless $patch_exe;
   _run_or_die("$patch_exe -f -s -p0 <$diff");
   unlink $diff or die "unlink $diff: $!\n";
 }
